refactor: Add local runtime interface
