+++ internal/deploy/diag_local_test.go
@@ -22,7 +22,10 @@ func TestDiagnoseLocalUnsafe_UnsupportedPlatform(t *testing.T) {
 	// test since Go runs all non-parallel tests to completion first).
 	deployment := newLocalTestDeployment(t)
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, nil))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, nil),
+	)
 
 	if diagnostics.PlatformSupported {
 		t.Fatalf("expected unsupported platform, got %+v", diagnostics)
@@ -45,7 +48,10 @@ description: test infrastructure
 backend: tofu
 `)
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, nil))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, nil),
+	)
 
 	if !diagnostics.PlatformSupported {
 		t.Fatal("expected platform support to bypass via EXASOL_LOCAL_ALLOW_UNSUPPORTED_PLATFORM")
@@ -63,7 +69,10 @@ func TestDiagnoseLocalUnsafe_VMNotRunning(t *testing.T) {
 	ensureLocalRuntimeWorkDir(t, deployment)
 	manager := writeFakeCombinedRunner(t, `{"running":false}`, "")
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, manager))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, manager),
+	)
 
 	if diagnostics.VMRunning == nil || *diagnostics.VMRunning {
 		t.Fatalf("expected VMRunning to be false, got %+v", diagnostics)
@@ -87,7 +96,10 @@ func TestDiagnoseLocalUnsafe_VMRunningReportsPortsAndHealth(t *testing.T) {
 	manager := writeFakeCombinedRunner(t, `{"running":true}`, healthJSON)
 	writeFakeVMState(t, deployment, "192.168.64.5", 20022, 28563, 0)
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, manager))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, manager),
+	)
 
 	if diagnostics.VMRunning == nil || !*diagnostics.VMRunning {
 		t.Fatalf("expected VMRunning to be true, got %+v", diagnostics)
@@ -115,7 +127,10 @@ func TestDiagnoseLocalUnsafe_VMRunningMatchesRunningState_NoWarning(t *testing.T
 	manager := writeFakeCombinedRunner(t, `{"running":true}`, `{"ports":{}}`)
 	writeFakeWorkflowState(t, deployment, &config.WorkflowStateRunning{})
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, manager))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, manager),
+	)
 
 	if diagnostics.Warning != "" {
 		t.Fatalf("expected no warning when workflow state matches a running VM, got %q",
@@ -135,7 +150,10 @@ func TestDiagnoseLocalUnsafe_VMRunningButStateNotRunning_Warning(t *testing.T) {
 		InterruptedDuringOperation: "start",
 	})
 
-	diagnostics := diagnoseLocalUnsafe(context.Background(), localruntime.New(deployment, manager))
+	diagnostics := diagnoseLocalUnsafe(
+		context.Background(),
+		localruntime.NewMacVMRuntime(deployment, manager),
+	)
 
 	if diagnostics.Warning == "" {
 		t.Fatal("expected a warning when a VM is running but the workflow state doesn't expect one")
